Intégration de imgui pour monitorer le jeu Grosse modification sur le makefile Utiliser -DIMGUI_DEBUG pour activer imgui Suppression du make map qui va être déplacer dans un projet à par entière Ne pas oublier d'installer le packet libbox2d-dev sur les distrib debian

+++ src/game/game.cpp
@@ -11,8 +11,8 @@ using namespace std;
 
 GLFWwindow* gameWindow = nullptr;
 
-float DELTA_TIME = 0.f;
-float FPS = 0.f; 
+double DELTA_TIME = 0;
+double FPS = 0;
 float GRAVITY = 9.8f;
 default_random_engine RANDOM_ENGINE;
